features/requirements.txt
behave
jpath
nose
PyMySQL
PyYAML
requests
trafaret
